+++ app/exceptions/navigation_log_errors.ts
@@ -2,6 +2,10 @@ export class NavigationLogNotFoundError extends Error {
   name = 'NavigationLogNotFoundError'
 }
 
+export class NavigationLogConflictError extends Error {
+  name = 'NavigationLogConflictError'
+}
+
 export class NavigationLogValidationError extends Error {
   name = 'NavigationLogValidationError'
   constructor(